+++ ruthless-plugin/src/main/java/dev/lajoscseppento/ruthless/plugin/impl/RuthlessExtensionImpl.java
@@ -1,7 +1,7 @@
 package dev.lajoscseppento.ruthless.plugin.impl;
 
 import dev.lajoscseppento.ruthless.plugin.RuthlessExtension;
-import dev.lajoscseppento.ruthless.plugin.logging.impl.RuthlessLogger;
+import dev.lajoscseppento.ruthless.plugin.logging.RuthlessLogger;
 import lombok.NonNull;
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.Configuration;