+++ ruthless-plugin/src/main/java/dev/lajoscseppento/ruthless/plugin/impl/RuthlessSettingsPlugin.java
@@ -2,9 +2,11 @@ package dev.lajoscseppento.ruthless.plugin.impl;
 
 import dev.lajoscseppento.gradle.plugin.common.GradleVersion;
 import dev.lajoscseppento.ruthless.plugin.configuration.impl.RuthlessConfiguration;
+import dev.lajoscseppento.ruthless.plugin.logging.RuthlessLoggingPlugin;
 import lombok.NonNull;
 import org.gradle.api.Plugin;
 import org.gradle.api.initialization.Settings;
+import org.gradle.api.invocation.Gradle;
 
 public class RuthlessSettingsPlugin implements Plugin<Settings> {
   @Override
@@ -12,10 +14,10 @@ public class RuthlessSettingsPlugin implements Plugin<Settings> {
     GradleVersion.of(settings)
         .requireAtLeast(RuthlessConfiguration.INSTANCE.getMinimumGradleVersion());
 
-    settings.getGradle().getPluginManager().apply(RuthlessGradlePlugin.class);
+    settings.getPluginManager().apply(RuthlessLoggingPlugin.class);
 
-    settings
-        .getGradle()
-        .allprojects(project -> project.getPluginManager().apply(RuthlessBasePlugin.class));
+    Gradle gradle = settings.getGradle();
+    gradle.getPluginManager().apply(RuthlessGradlePlugin.class);
+    gradle.allprojects(project -> project.getPluginManager().apply(RuthlessBasePlugin.class));
   }
 }